pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-server-odoo"
version = "0.6.0"
description = "A Model Context Protocol server for Odoo ERP systems"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Andrey Ivanov", email = "ivnv.xd@gmail.com" }
]
maintainers = [
    { name = "Andrey Ivanov", email = "ivnv.xd@gmail.com" }
]
license = { text = "MPL-2.0" }
keywords = ["mcp", "odoo", "erp", "model-context-protocol", "ai", "llm", "claude", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.26.0,<2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.urls]
Homepage = "https://github.com/ivnvxd/mcp-server-odoo"
Documentation = "https://github.com/ivnvxd/mcp-server-odoo#readme"
Repository = "https://github.com/ivnvxd/mcp-server-odoo"
Issues = "https://github.com/ivnvxd/mcp-server-odoo/issues"
Changelog = "https://github.com/ivnvxd/mcp-server-odoo/releases"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "ty>=0.0.18",
    "ruff>=0.15.2",
    "requests>=2.32.3",
]

[project.scripts]
mcp-server-odoo = "mcp_server_odoo.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server_odoo"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
markers = [
    "asyncio: marks tests as async (managed by pytest-asyncio)",
    "yolo: needs running Odoo instance (vanilla XML-RPC, no MCP module)",
    "mcp: needs running Odoo with MCP module installed",
]
filterwarnings = [
    "ignore:coroutine .* was never awaited:RuntimeWarning",
]

[tool.coverage.run]
source = ["mcp_server_odoo"]
omit = ["tests/*", "*/__pycache__/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.ty.environment]
python-version = "3.10"

[tool.ty.src]
include = ["mcp_server_odoo/"]
exclude = ["venv/", ".venv/", "build/", "dist/"]

[tool.ty.rules]
possibly-unresolved-reference = "warn"
invalid-argument-type = "ignore"
invalid-parameter-default = "error"
invalid-return-type = "warn"
invalid-assignment = "warn"
unresolved-attribute = "ignore"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
